remove legacy auth
